+++ src/main/java/com/vicious/persist/annotations/Save.java
@@ -14,7 +14,7 @@ import java.lang.annotation.Target;
  * @author Jack Andersen
  */
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.FIELD,ElementType.METHOD})
+@Target({ElementType.FIELD,ElementType.METHOD, ElementType.PARAMETER})
 public @interface Save {
     /**
      * The name of the savable element. If left empty this defaults to the Method or Field name.
@@ -40,5 +40,18 @@ public @interface Save {
          * @return The savable element's effective name.
          */
         String value();
+    }
+    /**
+     * A special sub-annotation that marks a Constructor as an initializer.
+     * The initializer should contain all {@link Save} fields as parameters.
+     * @since 1.4.2
+     */
+    @Retention(RetentionPolicy.RUNTIME)
+    @Target(ElementType.CONSTRUCTOR)
+    @interface Constructor{
+        /**
+         * @return An array containing all the names of the {@link Save} fields each parameter corresponds to.
+         */
+        String[] value();
     }
 }